pyproject.toml
[project]
name = "pref-uncertainty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "hf-xet>=1.1.5",
    "hydra-core>=1.3.2",
    "pandas>=2.3.0",
    "torch>=2.8.0",
    "transformers>=4.53.1",
    "mypy>=1.17.0",
    "google-genai>=1.28.0",
    "huggingface-hub>=0.33.4",
    "python-dotenv>=1.1.1",
    "wandb>=0.21.0",
    "peft>=0.17.0",
    "tiktoken>=0.9.0",
    "hf-transfer",
    "ipykernel>=7.0.1",
    "numpy>=2.2.6",
    "trl>=0.25.1",
    "pydantic>=2.12.4",
    "flash-attn",
    "openai",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.8",
    "scipy>=1.17.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
flash-attn = { url = "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.5.4/flash_attn-2.8.3+cu128torch2.9-cp312-cp312-linux_x86_64.whl" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff.lint]
extend-select = ["B"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]
